+++ webserver/src/main/java/cn/qkl/webserver/vo/cockpit/integrated/LegendNumberVO.java
@@ -22,5 +22,5 @@ public class LegendNumberVO {
     String name;
 
     @ApiModelProperty("数字")
-    Long number;
+    Integer number;
 }
